Improve the build system by decreasing the external dependencies used, and turn split-objs on by default.
This commit modifies the build system to download and install local copies of
most of the major build dependencies required by the HaLVM. This approach seemed
worthwhile based upon recent support requests, in which minor changes in the
versions of some of the dependencies (Cabal, for example) caused some really
weird problems. This way, we can fix everything.

As a side effect of testing this branch, this commit also merges in some changes
to underlying repositories (halvm-ghc and openlibm) that appear to fix the HaLVM
on i696/PAE.

(This is a squashed commit of the 'lessdep' branch. See that branch for more
detailed patches and information.)