Fix various XenStore issues
I suspect this fixes #39 and #40. Well, it does fix them, but I
suspect it does this without breaking a whole lot else.

Occasionally a XenStore event arrives somewhat slowly on the wire, and
our event handler fires multiple times. Sometimes this means the
parser fails after consuming some input, which means the event will
never parse correctly, and the rest of it gets discarded over the next
couple of iterations of `onXenbusEvent`. This leaves a Haskell thread
hung waiting on an MVar that isn't going to be replaced any time soon.

It could be problematic to let things buffer indefinitely whever there
is a no-parse. If some unparsable junk arrives on the wire, this
commit will break the XenStore for the remaining lifetime of the HaLVM
instance. That should probably be fixed by some careful person