Support packaging for a wider variety of Linux systems.
This patch should support automatic packaging (via 'make packages') for Fedora,
Ubuntu, and CentOS. It probably supports other Debian packaging systems, as
well, but that hasn't been tested. In addition, this should provide a fairly
obvious place to detect other systems and integrate automated packaging commands
for them.

This is a squashed commit of the 'ubuntu' branch.