Initial Amazon EC2 support.
This commit should allow HaLVM images to run under EC2. Any problems users face
running on HaLVMs should now be considered a bug. To make this happen, we
needed to backport early grant reference code, as EC2 has not been updated to
support the "version 2" form of said structures.

In order to make testing a bit easier, we also added the "PacketCount"
example to test networking without a full network stack.

(This is a squashed merge of the 'ec2' branch.)